WIP per file validation
